Update README.md
if you use the :0 display in the script for autostart you seem to run into an issue where vncserver is already running. Specifically, if you check the status of the service you will see: 

Warning: raspberrypi:0 is taken because of /tmp/.X0-lock
Remove this file if there is no X server raspberrypi:0
A VNC server is already running as :0

by using ":1" you avoid this issue and status looks clean.
thanks!